further refinement to User model, fixed spelling mistake, Added reset user password and update user password. removed these from user service and into auth service
